requirements.txt
asgiref==3.4.1
Django==4.0
django-bootstrap-static==4.2.1
django-inline-svg==0.1.1
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
Pillow==8.4.0
python-decouple==3.5
python-utils==3.1.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
Unidecode==1.3.2
